+++ RegionService/src/main/java/org/energygrid/east/regionservice/service/IRegionService.java
@@ -1,6 +1,7 @@
 package org.energygrid.east.regionservice.service;
 
 import org.energygrid.east.regionservice.model.House;
+import org.springframework.data.domain.Pageable;
 
 import java.util.List;
 
@@ -11,19 +12,20 @@ public interface IRegionService {
      * @param regionName is the name of the region
      * @return returns all info from the region
      */
-    List<House> getAllHousesProvince(String regionName,int page);
+    List<House> getAllHousesProvince(String regionName,long page);
 
 
     /**
      * @param cityName name of the city
+     * @param page
      * @return all houses in a city
      */
-    List<House> getAllHousesCity(String cityName,int page);
+    List<House> getAllHousesCity(String cityName, long page);
 
 
     /**
      * @param streetName name of the street
      * @return all houses in a street
      */
-    List<House> getAllHousesStreet(String streetName,int page);
+    List<House> getAllHousesStreet(String streetName,long page);
 }
